Rewrite console testFramework
